requirements.txt
asgiref==3.8.1
coolname==2.2.0
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
psycopg==3.2.4
sqlparse==0.5.3
tzdata==2025.1
